Deduplicate dataset_api
